Renamed file to be specifically for related stories
