+++ functions.php
@@ -17,12 +17,12 @@
  **/
 
 
-require_once( 'functions/base.php' );    # Base theme functions
-require_once( 'functions/admin.php' );   # Admin/login functions
-require_once( 'custom-taxonomies.php' ); # Where taxonomies are defined
-require_once( 'custom-post-types.php' ); # Where post types are defined
-require_once( 'functions/config.php' );  # Where site-level configuration settings are defined
-require_once( 'functions/acf-fields.php' );
+require_once( 'functions/base.php' );            # Base theme functions
+require_once( 'functions/admin.php' );           # Admin/login functions
+require_once( 'custom-taxonomies.php' );         # Where taxonomies are defined
+require_once( 'custom-post-types.php' );         # Where post types are defined
+require_once( 'functions/config.php' );          # Where site-level configuration settings are defined
+require_once( 'functions/related-stories.php' ); # Where related stories ACF fields and functions are defined
 
 
 // Plugin extras/overrides
